store middleware checks if the coin being added already exists in local storage and the portfolio container renders the portfolio saved in local storage

+++ lib/store.ts
@@ -6,7 +6,7 @@ import {
 import coinReducer from "@/lib/coinSlice";
 import currencyReducer from "@/lib/currencySlice";
 import { cryptoApi } from "./cryptoApi";
-import portfolioReducer, { addCoin } from "@/lib/portfolioSlice";
+import portfolioReducer, { addCoin, coin } from "@/lib/portfolioSlice";
 
 export const listenerMiddleware = createListenerMiddleware();
 
@@ -20,6 +20,15 @@ listenerMiddleware.startListening.withTypes<RootState, AppDispatch>()({
 
     if (portfolioExists) {
       const parsedPortfolio = JSON.parse(portfolioExists);
+      const coinExists = parsedPortfolio.coins.some((coin: coin) =>
+        state.coins.find(
+          () => state.coins[state.coins.length - 1].id === coin.id
+        )
+      );
+      if (coinExists) {
+        return;
+      }
+
       parsedPortfolio.coins.push(state.coins[state.coins.length - 1]);
       localStorage.setItem("portfolio", JSON.stringify(parsedPortfolio));
     } else {
